fix: Add Divider() to bottom of navigation bar
It bugged me whenever I opened the app that the icons in the bottom main
tab bar weren't centered and were too close to the bottom of the screen.
I added a divider to the bottom of the tab bar to center it and give a
bit more space from the bottom of the screen (or window in macOS). In my
personal opinion it makes it look much better.