feat: Blur profile pics if not following with some exceptions
Damus blurs images from people you don't follow but profile pics are a
huge loophole. None of them are blurred. In the NoteContentView, if the
image is blurred than the profile pic should be blurred as well.

I did make some exceptions. Obviously the users profile pic is never
blurred but it's also not blurred in the profile view. This could be
subject to change but my reasoning was that the profile pic should be
blurred when just scrolling and the event isn't from a friend, but if
you choose to click than all bets are off.