Adding MutelistManager.swift
This patch adds MutelistManager which contains separate sets for each type of muted item.

Whenever we receive a new event we parse it into those sets.

When checking to see if an event is muted, we simply check in each of those sets if it contains the given mute. For words/phrases we have to loop through each item in the set to see if the event contains the given word.

In future patches I will be implementing and integrating this new MutelistManager.

Lighting Address: fishcharlie@strike.me

Signed-off-by: Charlie Fish <contact@charlie.fish>
Signed-off-by: William Casarin <jb55@jb55.com>